Added login and logout authentication, Adjusted handlebars with classes for custom css styling.
Added login and logout authentication to the authController, dashboard, home and post controllers.  There was a bug where the logout button was not showing up and the user could login infinitely.  Adjusted this by passing the state through to each controller.

Adjusted handlebars with classes for custom css styling.  Added style.css for some minimalistic styling.